requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-genai==1.10.0
nest-asyncio==1.6.0
gunicorn==21.2.0
